+++ CHANGELOG.md
@@ -10,6 +10,7 @@ v2.3.0
 * Methods unimplemented on the server throw an exception only allowed by the RemoteFailure
   annotation. Before, UnimplementedException would be thrown always.
 * RemoteException now includes the remote address.
+* Added a lenient option for restorable methods.
 
 v2.2.0 (2023-02-04)
 ------